feat: User 정보 추가로 인한 수정

+++ backend/src/main/java/com/howmuch/backend/service/CommentService.java
@@ -34,6 +34,7 @@ public class CommentService {
                 .map(comment -> new CommentResponseDTO(
                         comment.getCommentId(),
                         comment.getPost().getPostId(),
+                        comment.getUser().getUserId(),
                         comment.getUser().getNickname(), // 유저 닉네임 가져오기
                         comment.getContent(),
                         comment.getCreatedAt(),
@@ -47,6 +48,7 @@ public class CommentService {
                 .map(comment -> new CommentResponseDTO(
                         comment.getCommentId(),
                         comment.getPost().getPostId(),
+                        comment.getUser().getUserId(),
                         comment.getUser().getNickname(),
                         comment.getContent(),
                         comment.getCreatedAt(),
@@ -71,6 +73,7 @@ public class CommentService {
         return new CommentResponseDTO(
                 createdComment.getCommentId(),
                 createdComment.getPost().getPostId(),
+                createdComment.getUser().getUserId(),
                 createdComment.getUser().getNickname(), // 유저 닉네임 가져오기
                 createdComment.getContent(),
                 createdComment.getCreatedAt(),
@@ -88,7 +91,13 @@ public class CommentService {
     }
 
     @Transactional
-    public void deleteComment(Long commentId) {
+    public void deleteComment(Long commentId, Long sessionUserId) {
+        PostComment comment = commentRepository.findById(commentId)
+                        .orElseThrow(() -> new IllegalArgumentException("댓글이 존재하지 않습니다"));
+
+        if(!comment.getUser().getUserId().equals(sessionUserId)) {
+            throw new IllegalArgumentException("삭제 권한이 없습니다");
+        }
         commentRepository.deleteById(commentId);
     }
